% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_spacetime.R
\name{.partition_add_data}
\alias{.partition_add_data}
\title{Add observations to an existing partition}
\usage{
.partition_add_data(p, locs, z)
}
\arguments{
\item{p}{existing partition as created with stmra_partition}

\item{locs}{a matrix with data locations (points are rows, coordinates columns) with named(!) columns "x", "y", and "t"}

\item{z}{numeric vector with data values and length identical to \code{nrows(locs)}}
}
\description{
Add observations to an existing partition
}
