% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_spacetime.R
\name{.partition_add_pred}
\alias{.partition_add_pred}
\title{Add prediction locations to an existing partition}
\usage{
.partition_add_pred(p, locs)
}
\arguments{
\item{p}{existing partition as created with stmra_partition}

\item{locs}{a matrix with prediction locations (points are rows, coordinates columns) with named(!) columns "x", "y", and "t"}
}
\description{
Add prediction locations to an existing partition
}
