% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_spacetime.R
\name{partition_spacetime}
\alias{partition_spacetime}
\title{Create a spaciotemporal recursive partition of a given domain of interest, and assign
knots, data, and prediction locations to regions}
\usage{
partition_spacetime(M, r = ceiling(length(z)/8^M), domain, locs, z,
  pred_locs = NULL, region_minsize = c(0, 0, 0), resolution = c(NA,
  NA, NA))
}
\arguments{
\item{M}{length-one integer, number of partitioning levels}

\item{r}{length-one integer, number of knots in regions above the last partitioning level}

\item{domain}{numeric length-6 vector including named(!) elements "xmin", "xmax", "ymin", "ymax", "tmin", "tmax"}

\item{locs}{a matrix with data locations (points are rows, coordinates columns) with named(!) columns "x", "y", and "t"}

\item{z}{numeric vector with data values and length identical to \code{nrows(locs)}}

\item{pred_locs}{matrix of prediction locations with named(!) colums "x", "y", and "t", or \code{NULL} for parameter estimation}

\item{region_minsize}{length-3 vector with minimum region size in x, y, and t dimensions (in this order)}

\item{resolution}{size of pixels in x, y, t dimensions (in this order), used to apply subpixel shifts while placing knots at upper partitioning levels}
}
\description{
Create a spaciotemporal recursive partition of a given domain of interest, and assign
knots, data, and prediction locations to regions
}
