% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRA_spacetime.R
\name{plot.stmra_prediction}
\alias{plot.stmra_prediction}
\title{Plot stmra predictions}
\usage{
\method{plot}{stmra_prediction}(x, ..., variance = FALSE)
}
\arguments{
\item{x}{Prediction as returned by \link{predict}}

\item{...}{Further plot arguments passed to raster::plot}

\item{variance}{logical; if TRUE, plot predicted variances instead of means}
}
\description{
Plot stmra predictions
}
