% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRA_spacetime.R
\name{stmra}
\alias{stmra}
\title{Estimate covariance function parameters of a spatiotemporal multiresolution approximation}
\usage{
stmra(partition, cov_fun, data, theta0, lower_bounds = NULL,
  upper_bounds = NULL, trace = FALSE, control = list(),
  optim.method = 1, ineq_constraints = NULL)
}
\arguments{
\item{partition}{A stmra partition as returned from \code{\link{stmra_partition}}}

\item{cov_fun}{A valid spatiotemporal covariance function}

\item{data}{A RasterBrick or RasterStack object or a matrix with data used for parameter estimation}

\item{theta0}{initial values}

\item{lower_bounds}{lower boundaries of parameter values}

\item{upper_bounds}{upper boundaries of parameter values}

\item{trace}{logical; print current parameter values and objective function value after each iteration}

\item{control}{list to adjust optimization parameters e.g. maximum number of iterations, or tolerance (see Details)}

\item{optim.method}{optimization method (1 = nloptr::bobyqa, 2 = nloptr::lbfgs, 3 = stats::constrOptim), see Details}

\item{ineq_constraints}{extended matrix with linear inequality constraints (rightmost column is RHS vector of the system)}
}
\value{
model with estimated covariance function parameters theta
}
\description{
Estimate covariance function parameters of a spatiotemporal multiresolution approximation
}
\details{
The supported options and ther names in the control argument vary with the used optimization method. Please check the specific method for details.
}
