% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRA_spacetime.R
\name{stmra_assess}
\alias{stmra_assess}
\title{Compare predictions with true observations}
\usage{
stmra_assess(pred, truth)
}
\arguments{
\item{pred}{Predictions as returned from \code{\link{predict.stmra}}}

\item{truth}{True observations as RasterStack / RasterBrick}
}
\value{
List with computed prediction scores and residuals
}
\description{
Compare predictions with true observations
}
