% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{stmra_cov_metric_exp}
\alias{stmra_cov_metric_exp}
\title{Metric exponential spatiotemporal covariance function}
\usage{
stmra_cov_metric_exp(locs1, locs2, theta)
}
\arguments{
\item{locs1}{matrix with rows representing points and columns representing coordinates}

\item{locs2}{matrix with rows representing points and columns representing coordinates}

\item{theta}{vector of parameters (see Details)}
}
\description{
Compute pairwise covariances between two sets of locations using
a metric exponential spatiotemporal model.
}
\details{
Parameter vector \code{theta} includes range, sill, nugget, spatiotemporal anisotropy (in this order).
}
