% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{stmra_cov_nonstationary_R2_separable}
\alias{stmra_cov_nonstationary_R2_separable}
\title{Nonstationary spatiotemporal covariance function varying by latitude}
\usage{
stmra_cov_nonstationary_R2_separable(locs1, locs2, theta)
}
\arguments{
\item{locs1}{matrix with rows representing points and columns representing coordinates}

\item{locs2}{matrix with rows representing points and columns representing coordinates}

\item{theta}{vector of parameters (see Details)}
}
\description{
Compute pairwise covariances between two sets of locations using
a nonstationary kernel convolution approach with standard deviation and anisotropy varying as weighted sum of
Gaussian radial basis functions cenetered at different latitudes.
}
\details{
Parameter vector \code{theta} includes spatial nugget, temporal nugget, temporal range, standard deviation by latitude (9 values), spatial range in east-west direction (9 values), spatial range in south-north direction (9 values) (in this order).
}
\references{
Paciorek, C. J., & Schervish, M. J. (2006). Spatial modelling using a new class of nonstationary covariance functions. Environmetrics: The official journal of the International Environmetrics Society, 17(5), 483-506.
}
