% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{stmra_cov_porcu_etal_15}
\alias{stmra_cov_porcu_etal_15}
\title{Non-separable spatiotemporal covariance function on spherical distances}
\usage{
stmra_cov_porcu_etal_15(locs1, locs2, theta)
}
\arguments{
\item{locs1}{matrix with rows representing points and columns representing coordinates}

\item{locs2}{matrix with rows representing points and columns representing coordinates}

\item{theta}{vector of parameters (see Details)}
}
\description{
Compute pairwise covariances between two sets of locations using
a non-separable spatiotemporal model with Great circle distances from Porcu et al. (2016), eq. (15).
}
\details{
Parameter vector \code{theta} includes partial sill, spatial scale, temporal scale, nugget (in this order).
}
\references{
Porcu, E., Bevilacqua, M., & Genton, M. G. (2016). Spatio-temporal covariance and cross-covariance functions of the great circle distance on a sphere. Journal of the American Statistical Association, 111(514), 888-898.
}
