% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRA_spacetime.R
\name{stmra_create_model}
\alias{stmra_create_model}
\title{Create and initialize a spatiotemporal multiresolution approximation model}
\usage{
stmra_create_model(data, M, r, cov_fun, region_minsize = c(0, 0, 0))
}
\arguments{
\item{data}{A RasterBrick or RasterStack object with data used for parameter estimation}

\item{M}{length-one integer, number of partitioning levels}

\item{r}{length-one integer, number of knots in regions above the last partitioning level}

\item{region_minsize}{length-3 vector with minimum region size in x, y, and t dimensions (in this order)}

\item{covariance}{A valid spatiotemporal covariance function}
}
\value{
A stmra object as a list with elements  TODO
}
\description{
Create and initialize a spatiotemporal multiresolution approximation model
}
