% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRA_spacetime.R
\name{stmra_estimate}
\alias{stmra_estimate}
\title{Estimate covariance function parameters of a spatiotemporal multiresolution approximation}
\usage{
stmra_estimate(model, theta0, lower_bounds = NULL, upper_bounds = NULL,
  trace = FALSE, max.iter = NULL, ftol.abs = NULL,
  optim.method = 1, ineq_constraints = NULL)
}
\arguments{
\item{model}{A stmra model as returned from \code{\link{stmra_create_model}}}

\item{theta0}{initial values}

\item{lower_bounds}{lower boundaries of parameter values}

\item{upper_bounds}{upper boundaries of parameter values}

\item{trace}{logical; print current parameter values and objective function value after each iteration}

\item{max.iter}{maximum number of iterations in numerical optimization}

\item{ftol.abs}{absolute tolerance in numerical optimization}

\item{optim.method}{optimization method (1 = BOBYQA from nloptr package, 2 = LL-BFGS-B, 3 = constrOptim from constrOptim package), see details}

\item{ineq_constraints}{extended matrix with linear inequality constraints (rightmost column is RHS vector of the system)}
}
\value{
model with estimated covariance function parameters theta
}
\description{
Estimate covariance function parameters of a spatiotemporal multiresolution approximation
}
\details{
For optimization method 3 (constrOptim), \code{max.iter} can be a vector of length two, representing the maximum number of outer and inner iterations
}
