% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_spacetime.R
\name{stmra_partition}
\alias{stmra_partition}
\title{Create a spaciotemporal recursive partition of a given domain of interest, and assign
knots, data, and prediction locations to regions}
\usage{
stmra_partition(M, r, domain, region_minsize = c(0, 0, 0),
  resolution = c(NA, NA, NA))
}
\arguments{
\item{M}{length-one integer, number of partitioning levels}

\item{r}{length-one integer, number of knots in regions above the last partitioning level}

\item{domain}{numeric length-6 vector including named(!) elements "xmin", "xmax", "ymin", "ymax", "tmin", "tmax", or RasterStack / RasterBrick}

\item{region_minsize}{length-3 vector with minimum region size in x, y, and t dimensions (in this order)}

\item{resolution}{size of pixels in x, y, t dimensions (in this order) or RasterStack / RasterBrick object, used to apply subpixel shifts while placing knots at upper partitioning levels.}
}
\description{
Create a spaciotemporal recursive partition of a given domain of interest, and assign
knots, data, and prediction locations to regions
}
