% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRA_spacetime.R
\name{stmra_stack_to_matrix}
\alias{stmra_stack_to_matrix}
\title{Convert a raster stack to a data matrix}
\usage{
stmra_stack_to_matrix(s)
}
\arguments{
\item{s}{A RasterStack or RasterBrick object where layer dimension corresponds to time}
}
\value{
A data.frame with rows representing pixels and columns for coordinates ("x","y","t") and data values ("values")
}
\description{
Convert a raster stack to a data matrix
}
\note{
This functions assumes equidistant images, i.e. a constant temporal gap between consecutive layers
}
